OSX: Sandboxed basedir selection and raw input.
Switch raw input to GameController which is made to work with
sandboxed applications.

When the application starts, the user is prompted with a file
dialog that allows selection of what basedir to use. This will
be stored in the application preferences for future launches.

When the directory becomes unavailable, or a new version is
installed, the user will be prompted with the file dialog again.