RENDERER: Add support for skywind.
Feature available in the SP community. Ported mostly verbatim from
the IronWail engine. Partial cubemap alpha which gets blended with
itself to create the illusion of a moving sky. Quite hypnotizing.

To simplify tuning of the skybox the following commands are included
in the port:

* /skywind
* /skywind_save
* /skywind_load
* /skywind_lookdir
* /skywind_rotate

If a config file named $skyboxname_wind.cfg in the same directory as
the skybox is found, it's automatically loaded and configures the sky.

Implemented for both modern and classic renderers.