ENTS: Restore aliasmodel drawcalls limit to 512.
Even if the protocol supports more entities, lets pretend that
everything is as it was until this no longer results in slowdowns.

In practice it's harmless as entities are spread across both bmodels
and aliasmodels, and bmodels are what is expected to grow with more
advanced maps.

Should the limit be reached, there will be a warning, there are bounds
check on use, so it will not crash.