Add color 14, 15 and 16.
There was a bug in the following client versions that allowed players to
use more than the regular colors.

- qwcl-1.50
- qwcl-1.55
- qwcl-1.64
- qwcl-2.00
- qwcl-2.10
- glqwcl-1.64
- glqwcl-2.10

The bug was fixed in:

- qwcl-2.20
- glqwcl-2.20

Both the software and GL clients correctly restricted colors to the
range of 0 - 13 when using the color command. However, the developers
forgot to apply this restriction to the bottomcolor and topcolor cvars.

The behavior differed slightly between the qwcl and glqwcl clients. In
the software version, only two additional colors were available: color
14 (orange) and color 15 (bright red). In the GL version, players could
select from the entire palette.

This commit aims to reintroduce the most popular colors used during that
time:

- Orange (color 14, mapped to palette index 232)
- Bright red (color 15, mapped to palette index 248)
- Black (color 16, mapped to palette index 0)